service: Refactor internal type for de-duplication
This commit refactors the `svcInfo` internal type to the service
package.

As we add more fields to the loadbalancer.SVC struct, it becomes tedious
to add the same field in `svcInfo` simply to pass information down the
call stack. Upon inspection, the `svcInfo` struct is almost a mirror to
the loadbalancer.SVC struct, minus a few fields. Therefore, there's no
reason to keep them separate if they are sharing the same data anyway,
and we might as well embed it to ease future modifications.

Signed-off-by: Chris Tarazi <chris@isovalent.com>