lbmap: Store Maglev table size inside lb{4,6}_services
We store the Maglev table size inside the backend ID field which is
uint32, while the Maglev table size itself is uint64. The maximum
possible table size that we currently support is 131071, which easily
fits into a uint32, so downgrade the integer is OK.

The table size is stored inside the cilium_lb{4,6}_services_v2 as a map
value from key with backend_slot=0. This is done in order for the
datapath to fetch backends from the Maglev LUT (lookup table) without
overflowing on the size, since services can now have differing arrays
storing the backends.

Signed-off-by: Chris Tarazi <chris@isovalent.com>