watchers: Add callbacks implementation
This commit implements a package called callbacks for the K8s resource /
object watchers. It contains a list that holds callbacks for handling an
add, update, or delete event, respectively.

This allows the watchers to be simpler and leaner than they previously
were. They no longer need to hold logic specific to the consumers.
Instead, they convert the object (since it's an `interface{}`) and loops
over the callbacks in the list, calling them one by one.

In order for the consumers to be "notified" of events, they must
implement the k8s.io/client-go/tools/cache.ResourceEventHandler
interface and define the methods. In addition, they must
(*List).Register() themselves, so that the watcher is aware of them.

Signed-off-by: Chris Tarazi <chris@isovalent.com>