endpoint, policy: Break up policy preparation from implementation
This commit breaks up "policy calculation" into two steps:

1. Preparation
2. Implementation

Step (1) prepares the policy engine for the changes needed to implement
a policy revision. Specifically, the endpoint's `selectorPolicy` is
prepared.

Step (2) carries out the changes prepared in (1) to implement a revision
(represented by the state stored in `e.selectorPolicy`). The
"implementation" refers to the consumption of the `selectorPolicy`, i.e.
`e.selectorPolicy.Consume()`.

This commit allows step (1) to be done without the main endpoint lock
(`e.mutex`) held. It is crucial that the endpoint lock is not held for
subsequent commits which will allow the SelectorCache to potentially
call back into the ipcache for CIDR identity allocation and updating of
policymaps in the datapath. The latter includes endpoint regeneration,
which would cause a deadlock if step (1) held the endpoint lock.

Signed-off-by: Chris Tarazi <chris@isovalent.com>