endpoint: convert realizedRedirects to a sync.Map
Right now, policy generation requres holding the endpoint lock
throughout, but that's not actually needed. As part of removing the
endpoint lock, we need to make access to the proxy port redirects
thread-safe.

This map is only updated with the endpoint lock held, but rather than
adding a new lock and risking deadlocks, let's just make it a lock-free
sync.Map.

Signed-off-by: Casey Callendrello <cdc@isovalent.com>